feat: Update DoctorAbout.jsx and formatDate.jsx
- Imported the formatDate function in the DoctorAbout component.
- Added experience info with appropriate HTML.
- Formatted date info using the imported formatDate function.

+++ frontend/src/utils/formateDate.jsx
@@ -1,9 +1,7 @@
-import React from 'react'
-
 export const formateDate = (date, config) => {
   const defaultOptions = {
     day: 'numeric',
-    month: 'long',
+    month: 'short',
     year: 'numeric',
   };
   const options = config? config: defaultOptions;
